requirements.txt
tkcalendar>=1.6.1
Pillow>=10.4.0